pagination and sort added

+++ models/filter.go
@@ -10,9 +10,9 @@ type BaseFilter struct {
 	ID        primitive.ObjectID `json:"id" bson:"_id"`
 	Sort      string             `json:"sort" query:"sort" bson:"sort"`
 	LastIndex string             `json:"lastIndex" query:"lastIndex" bson:"lastIndex"`
-	Page      int                `json:"page" query:"page" bson:"page"`
-	Limit     int                `json:"limit" query:"limit" bson:"limit"`
-	NextPage  int                `json:"nextPage" query:"nextPage" bson:"nextPage"`
+	Page      int64              `json:"page" query:"page" bson:"page"`
+	Limit     int64              `json:"limit" query:"limit" bson:"limit"`
+	NextPage  int64              `json:"nextPage" query:"nextPage" bson:"nextPage"`
 }
 
 //AdvertiseFilter
